Made check_out file a core.utils function
