+++ virtual_rainforest/core/config.py
@@ -40,6 +40,7 @@ from jsonschema import Draft202012Validator, FormatChecker, exceptions, validato
 
 from virtual_rainforest.core.exceptions import ConfigurationError
 from virtual_rainforest.core.logger import LOGGER
+from virtual_rainforest.core.utils import check_outfile
 
 if sys.version_info[:2] >= (3, 11):
     import tomllib
@@ -235,54 +236,6 @@ def check_dict_leaves(
                 conflicts.append("%s" % ".".join(path + [str(key)]))
 
     return conflicts
-
-
-def check_outfile(merge_file_path: Path) -> None:
-    """Check that final output file is not already in the output folder.
-
-    Args:
-        merge_file_path: Path to save merged config file to (i.e. folder location + file
-            name)
-
-    Raises:
-        ConfigurationError: If the final output directory doesn't exist, isn't a
-            directory, or the final output file already exists.
-    """
-
-    # Extract parent folder name and output file name. If this is a relative path, it is
-    # expected to be relative to where the command is being run.
-    if not merge_file_path.is_absolute():
-        parent_fold = merge_file_path.parent.relative_to(".")
-    else:
-        parent_fold = merge_file_path.parent
-    out_file_name = merge_file_path.name
-
-    # Throw critical error if the output folder doesn't exist
-    if not Path(parent_fold).exists():
-        to_raise = ConfigurationError(
-            f"The user specified output directory ({parent_fold}) doesn't exist!"
-        )
-        LOGGER.critical(to_raise)
-        raise to_raise
-
-    elif not Path(parent_fold).is_dir():
-        to_raise = ConfigurationError(
-            f"The user specified output folder ({parent_fold}) isn't a directory!"
-        )
-        LOGGER.critical(to_raise)
-        raise to_raise
-
-    # Throw critical error if combined output file already exists
-    if merge_file_path.exists():
-        to_raise = ConfigurationError(
-            f"A file in the user specified output folder ({parent_fold}) already "
-            f"makes use of the specified output file name ({out_file_name}), this "
-            f"file should either be renamed or deleted!"
-        )
-        LOGGER.critical(to_raise)
-        raise to_raise
-
-    return None
 
 
 def collect_files(cfg_paths: list[str]) -> list[Path]:
